+++ lib/src/app.dart
@@ -406,7 +406,7 @@ class _FluentAppState extends State<FluentApp> {
   Widget _builder(BuildContext context, Widget? child) {
     final themeData = theme(context);
     final mTheme = context.findAncestorWidgetOfExactType<m.Theme>();
-    return m.Theme(
+    return m.AnimatedTheme(
       data: mTheme?.data ?? m.ThemeData(brightness: themeData.brightness),
       child: AnimatedFluentTheme(
         curve: themeData.animationCurve,